Camera Ctrls & Character States
**RESOLVING SYNCING ISSUES**

Tweaked the Camera Angles and controls |
- Zoom In/Out:+1:
- Zoom power :+1:
** Implemented to be changed in editor **

WIP for character states for placeholder Animations

Step 1: Implement the code for the player input. :+1:
Step 2: Configure the animations to play when current state is changed.
:+1:
Step 3: Enable colliders on the player mesh to be enabled or disabled
for melee combat.
Step 4: Create Projectile class. :+1:
Step 5: Spawn actor when casting is triggered and conditions are met.
Step 6: Tweak animations to be playable for the build.
Step 7: Enable permissions to change the damage output for each attack
type. :+1:

Mixamo:tm: Animations placeholders until Character models become
available + animations. **USE FOR PROTOTYPE**